REFACTOR: Remove bitdubai from network service asset transmission plugin. #5802
