+++ DAP/plugin/network_service/fermat-dap-plugin-network-service-asset-transmission/src/main/java/org/fermat/fermat_dap_plugin/layer/network/service/asset/transmission/developer/bitdubai/version_2/database/communications/CommunicationNetworkServiceDatabaseFactory.java
@@ -4,7 +4,7 @@
   * You may not modify, use, reproduce or distribute this software.
  * BITDUBAI/CONFIDENTIAL
  */
-package com.bitdubai.fermat_dap_plugin.layer.network.service.asset.transmission.developer.bitdubai.version_2.database.communications;
+package org.fermat.fermat_dap_plugin.layer.network.service.asset.transmission.developer.bitdubai.version_2.database.communications;
 
 import com.bitdubai.fermat_api.layer.osa_android.database_system.Database;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.DatabaseDataType;
