+++ DAP/plugin/network_service/fermat-dap-plugin-network-service-asset-transmission/src/main/java/org/fermat/fermat_dap_plugin/layer/network/service/asset/transmission/developer/bitdubai/version_2/exceptions/CantUpdateRecordDataBaseException.java
@@ -1,16 +1,16 @@
 /*
- * @#CantDeleteRecordDataBaseException.java - 2015
+ * @#CantUpdateRecordDataBaseException.java - 2015
  * Copyright bitDubai.com., All rights reserved.
  * You may not modify, use, reproduce or distribute this software.
  * BITDUBAI/CONFIDENTIAL
  */
-package com.bitdubai.fermat_dap_plugin.layer.network.service.asset.transmission.developer.bitdubai.version_2.exceptions;
+package org.fermat.fermat_dap_plugin.layer.network.service.asset.transmission.developer.bitdubai.version_2.exceptions;
 
 
 import com.bitdubai.fermat_api.FermatException;
 
 /**
- * The Class <code>com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.user.developer.bitdubai.version_1.exceptions.CantDeleteRecordDataBaseException</code> is
+ * The Class <code>com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.user.developer.bitdubai.version_1.exceptions.CantUpdateRecordDataBaseException</code> is
  * throw when error occurred updating new record in a table of the data base
  * <p/>
  * Created by Hendry Rodriguez - (elnegroevaristo@gmail.com) on 07/10/15.
@@ -18,12 +18,12 @@ import com.bitdubai.fermat_api.FermatException;
  * @version 1.0
  * @since Java JDK 1.7
  */
-public class CantDeleteRecordDataBaseException extends FermatException {
+public class CantUpdateRecordDataBaseException extends FermatException {
 
     /**
      * Represent the default message
      */
-    public static final String DEFAULT_MESSAGE = "CAN'T DELETE RECORD ON DATABASE";
+    public static final String DEFAULT_MESSAGE = "CAN'T UPDATE RECORD ON DATABASE";
 
     /**
      * Constructor with parameters
@@ -33,7 +33,7 @@ public class CantDeleteRecordDataBaseException extends FermatException {
      * @param context
      * @param possibleReason
      */
-    public CantDeleteRecordDataBaseException(final String message, final Exception cause, final String context, final String possibleReason) {
+    public CantUpdateRecordDataBaseException(final String message, final Exception cause, final String context, final String possibleReason) {
         super(message, cause, context, possibleReason);
     }
 
@@ -43,7 +43,7 @@ public class CantDeleteRecordDataBaseException extends FermatException {
      * @param message
      * @param cause
      */
-    public CantDeleteRecordDataBaseException(final String message, final Exception cause) {
+    public CantUpdateRecordDataBaseException(final String message, final Exception cause) {
         this(message, cause, "", "");
     }
 
@@ -52,7 +52,7 @@ public class CantDeleteRecordDataBaseException extends FermatException {
      *
      * @param message
      */
-    public CantDeleteRecordDataBaseException(final String message) {
+    public CantUpdateRecordDataBaseException(final String message) {
         this(message, null);
     }
 
@@ -61,7 +61,7 @@ public class CantDeleteRecordDataBaseException extends FermatException {
      *
      * @param exception
      */
-    public CantDeleteRecordDataBaseException(final Exception exception) {
+    public CantUpdateRecordDataBaseException(final Exception exception) {
         this(exception.getMessage());
         setStackTrace(exception.getStackTrace());
     }
@@ -69,7 +69,7 @@ public class CantDeleteRecordDataBaseException extends FermatException {
     /**
      * Constructor
      */
-    public CantDeleteRecordDataBaseException() {
+    public CantUpdateRecordDataBaseException() {
         this(DEFAULT_MESSAGE);
     }
 }
